Major reworking of the interpreter.
First attempt at addressing issues discussed in #136.

The main issue this addresses is that it's not possible
to reliably detect where a Redis server response ends
without actually performing RESP deserialization.

Currently the RESP serialization is mixed in with
the command-specific handling in Input.scala and
Output.scala. This commit splits out the RESP
serialization, which works the same for all commands,
into RespValue.scala, which constructs an ADT description
of the the RESP values.

This produced major changes in Output.scala, as the
command-specific processing is no longer examining raw
RESP strings, but is instead pattern matching on the
RespValue ADT.

Interpreter has been rewritten to accept and return
RespValues. It now uses AsynchronousSocketChannel, as
that's the simplest way to avoid busy-waiting on
a non-blocking socket, without needing blocking threads.
More work is likely needed to gracefully handle
IOExceptions.

The RESP deserialization is implemented in the most
straight-forward way, processing one byte at a time.
This will be slow, it will need to be replaced by
Chunk-based sinks to perform decently.

More tests of the RESP serialization/deserialization are
needed. It's nice that this can now be tested independently
of specific commands. Command tests that were
merely testing serialization have been removed.